requirements.txt
certifi==2018.4.16
chardet==3.0.4
future==0.16.0
idna==2.6
kafka-python==1.4.2
oauthlib==2.0.7
PySocks==1.6.8
python-dotenv==0.8.2
python-twitter==3.4.1
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
tweepy==3.6.0
urllib3==1.22
